Add tags controller

+++ src/main/java/com/example/demo/dto/TagWithQuantityDto.java
@@ -7,12 +7,12 @@ public class TagWithQuantityDto {
 
     private String name;
 
-    private Integer quantity;
+    private Long quantity;
 
     public TagWithQuantityDto() {
     }
 
-    public TagWithQuantityDto(Long id, String name, Integer quantity) {
+    public TagWithQuantityDto(Long id, String name, Long quantity) {
         this.id = id;
         this.name = name;
         this.quantity = quantity;
@@ -34,11 +34,11 @@ public class TagWithQuantityDto {
         this.name = name;
     }
 
-    public Integer getQuantity() {
+    public Long getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(Long quantity) {
         this.quantity = quantity;
     }
 }
